上传视频接口、上传视频缩略图接口、增加ffmpeg 工具类

+++ pojo/src/main/java/com/nekosighed/pojo/Dto/UploadVideoInfoDto.java
@@ -31,7 +31,7 @@ public class UploadVideoInfoDto {
      */
     @ApiModelProperty(value = "视频时长", name = "videoSeconds", example = "40.1", required = true)
     @NotNull(message = "视频时长不能为空")
-    private double videoSeconds;
+    private Float videoSeconds;
 
     /**
      * 视频长度
